fix: change budget type from Integer to Double in SignalementDto and update related logic

+++ src/main/java/com/route/FirebaseDTO/SignalementDto.java
@@ -8,7 +8,7 @@ public class SignalementDto {
     private String id;
     private String idFirestore;
     private Integer idSignalement;
-    private Integer budget;
+    private Double budget;
     private Integer avancement;
     private Integer surface;
